moved spawn files to own package
